added authentication for the rtmp publish and play streams
moved the file uploading and watching to its own module
