+++ package.json
@@ -31,7 +31,9 @@
     "express": "^4.16.3",
     "fluent-ffmpeg": "^2.1.2",
     "lodash": "^4.17.10",
+    "md5": "^2.2.1",
     "mkdirp": "^0.5.1",
+    "moment": "^2.24.0",
     "uuid": "^3.3.2",
     "ws": "^5.1.1"
   },
